requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
atlassian-python-api==2.1.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
dataclasses==0.6
Deprecated==1.2.10
discord==1.0.1
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.0
osrsbox==2.0.7
OSRSBytes==1.2.3
pipenv==2020.11.15
pylint==2.6.0
python-osrsapi==0.0.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
wrapt==1.12.1
yarl==1.6.3
